+++ apps/replays/src/index.ts
@@ -1,3 +1,4 @@
+import { AuthFactory } from 'auth'
 import cors from 'cors'
 import express from 'express'
 import { HTTPLogger } from 'metrics'
@@ -6,9 +7,15 @@ import { replayRouter } from './api/router'
 
 const app = express()
 const logger = new HTTPLogger('Replay')
+const { handler } = AuthFactory.supabase({
+  url: process.env.SUPABASE_URL!,
+  key: process.env.SUPABASE_KEY!,
+  secret: process.env.JWT_SECRET!,
+  issuer: process.env.JWT_ISSUER!,
+})
 
 setupMessageQueue()
 
 app.use(cors())
-replayRouter.mount(app, logger)
+replayRouter.mount(app, logger, handler)
 app.listen(process.env.REPLAYS_PORT)