+++ packages/api-schema/src/schema/evc/index.ts
@@ -2,6 +2,7 @@ import { decodeVersion } from '@ivy-chess/model'
 import { z } from 'zod'
 import { engineConfigSchema, engineNameSchema, engineVariationSchema } from '../../shared/engine'
 import { downloadSchema } from '../../shared/generic'
+import { contributorRoles } from '../../shared/user'
 import { endpoint } from '../../types/endpoint'
 import { route } from '../../types/route'
 
@@ -32,15 +33,22 @@ export const createSchema = z.object({
  * Schema for the engine version control API.
  */
 export const engineVersioningRoute = route('/engines', {
-  all: endpoint('/', 'GET').success(z.array(engineConfigSchema)),
-  create: endpoint('/', 'POST').body(createSchema).files(['engine']).success(engineConfigSchema),
+  all: endpoint('/', 'GET').unprotected().success(z.array(engineConfigSchema)),
+  create: endpoint('/', 'POST')
+    .access(...contributorRoles)
+    .body(createSchema)
+    .files(['engine'])
+    .success(engineConfigSchema),
   get: endpoint('/:id', 'GET')
+    .unprotected()
     .params(z.object({ id: z.string().nonempty() }))
     .success(engineConfigSchema),
   download: endpoint('/bin/:engine/:id', 'GET')
+    .unprotected()
     .params(z.object({ engine: engineNameSchema, id: z.string().nonempty() }))
     .success(downloadSchema),
   delete: endpoint('/:engine/:id', 'DELETE')
+    .access(...contributorRoles)
     .params(
       z.object({
         engine: engineNameSchema,
@@ -49,6 +57,7 @@ export const engineVersioningRoute = route('/engines', {
     )
     .success(z.object({ success: z.boolean() })),
   getVersion: endpoint('/:name/:version', 'GET')
+    .unprotected()
     .params(
       z.object({
         name: engineNameSchema,