+++ packages/api-schema/src/types/endpoint.ts
@@ -34,6 +34,7 @@ const defaultFailureSchema = z.object({
  * Creates a new endpoint with default input and output schemas.
  * The default input schema expects an empty query, an undefined body and empty params.
  * The default output schema expects an undefined success response and a failure response with a message property.
+ * An endpoint requires authentication by default.
  *
  * @example
  * ```ts
@@ -71,13 +72,15 @@ export function endpoint(path: string, method: Method) {
       success: z.undefined({ invalid_type_error: 'Expected undefined success response' }),
       failure: defaultFailureSchema,
     },
+    true,
+    new Set(),
   )
 }
 
 /**
  * An endpoint represents the schema of a single API endpoint.
  * It defines which input the endpoint accepts and which output it returns.
- * The input consists of query parameters, a body, attached files, and path parameters aka params.
+ * The input consists of query parameters, a body, attached files, authentication data and path parameters aka params.
  * Each input and output is defined by a Zod schema and can be validated at runtime
  * when data is received over an untrusted network.
  */
@@ -87,6 +90,7 @@ export class Endpoint<
   P extends z.ZodType,
   U extends string,
   S extends z.ZodType,
+  A extends boolean,
   F extends z.ZodType = typeof defaultFailureSchema,
 > {
   /**
@@ -101,6 +105,11 @@ export class Endpoint<
   public readonly method: Method
 
   /**
+   * Whether the endpoint requires authentication.
+   */
+  public readonly authenticated: A
+
+  /**
    * The input schema of the endpoint.
    */
   protected readonly input: Input<Q, B, P, U>
@@ -110,11 +119,43 @@ export class Endpoint<
    */
   protected readonly output: Output<S, F>
 
-  constructor(path: string, method: Method, input: Input<Q, B, P, U>, output: Output<S, F>) {
+  /**
+   * The accepted access roles of the endpoint.
+   */
+  protected readonly accessRoles: Set<string>
+
+  constructor(
+    path: string,
+    method: Method,
+    input: Input<Q, B, P, U>,
+    output: Output<S, F>,
+    authenticated: A,
+    accessRoles: Set<string>,
+  ) {
     this.path = path
     this.method = method
     this.input = input
     this.output = output
+    this.authenticated = authenticated
+    this.accessRoles = new Set(accessRoles)
+  }
+
+  /**
+   * Creates a new endpoint with the same path and method but requires authentication.
+   *
+   * @returns A new endpoint with updated authentication requirement.
+   */
+  public protected(): Endpoint<Q, B, P, U, S, true, F> {
+    return new Endpoint(this.path, this.method, this.input, this.output, true, this.accessRoles)
+  }
+
+  /**
+   * Creates a new endpoint with the same path and method but does not require authentication.
+   *
+   * @returns A new endpoint with updated authentication requirement.
+   */
+  public unprotected(): Endpoint<Q, B, P, U, S, false, F> {
+    return new Endpoint(this.path, this.method, this.input, this.output, false, this.accessRoles)
   }
 
   /**
@@ -123,8 +164,15 @@ export class Endpoint<
    * @param query The new query schema.
    * @returns A new endpoint with updated query schema.
    */
-  public query<T extends z.ZodType>(query: T): Endpoint<T, B, P, U, S, F> {
-    return new Endpoint(this.path, this.method, { ...this.input, query }, this.output)
+  public query<T extends z.ZodType>(query: T): Endpoint<T, B, P, U, S, A, F> {
+    return new Endpoint(
+      this.path,
+      this.method,
+      { ...this.input, query },
+      this.output,
+      this.authenticated,
+      this.accessRoles,
+    )
   }
 
   /**
@@ -133,8 +181,15 @@ export class Endpoint<
    * @param body The new body schema.
    * @returns A new endpoint with updated body schema.
    */
-  public body<T extends z.ZodType>(body: T): Endpoint<Q, T, P, U, S, F> {
-    return new Endpoint(this.path, this.method, { ...this.input, body }, this.output)
+  public body<T extends z.ZodType>(body: T): Endpoint<Q, T, P, U, S, A, F> {
+    return new Endpoint(
+      this.path,
+      this.method,
+      { ...this.input, body },
+      this.output,
+      this.authenticated,
+      this.accessRoles,
+    )
   }
 
   /**
@@ -143,8 +198,15 @@ export class Endpoint<
    * @param params The new params schema.
    * @returns A new endpoint with updated params schema.
    */
-  public params<T extends z.ZodType>(params: T): Endpoint<Q, B, T, U, S, F> {
-    return new Endpoint(this.path, this.method, { ...this.input, params }, this.output)
+  public params<T extends z.ZodType>(params: T): Endpoint<Q, B, T, U, S, A, F> {
+    return new Endpoint(
+      this.path,
+      this.method,
+      { ...this.input, params },
+      this.output,
+      this.authenticated,
+      this.accessRoles,
+    )
   }
 
   /**
@@ -153,8 +215,15 @@ export class Endpoint<
    * @param files The new files list of expected files.
    * @returns A new endpoint with updated files list.
    */
-  public files<T extends string>(files: T[]): Endpoint<Q, B, P, T, S, F> {
-    return new Endpoint(this.path, this.method, { ...this.input, files }, this.output)
+  public files<T extends string>(files: T[]): Endpoint<Q, B, P, T, S, A, F> {
+    return new Endpoint(
+      this.path,
+      this.method,
+      { ...this.input, files },
+      this.output,
+      this.authenticated,
+      this.accessRoles,
+    )
   }
 
   /**
@@ -163,8 +232,15 @@ export class Endpoint<
    * @param success The new success schema.
    * @returns A new endpoint with updated success schema.
    */
-  public success<T extends z.ZodType>(success: T): Endpoint<Q, B, P, U, T, F> {
-    return new Endpoint(this.path, this.method, this.input, { ...this.output, success })
+  public success<T extends z.ZodType>(success: T): Endpoint<Q, B, P, U, T, A, F> {
+    return new Endpoint(
+      this.path,
+      this.method,
+      this.input,
+      { ...this.output, success },
+      this.authenticated,
+      this.accessRoles,
+    )
   }
 
   /**
@@ -173,8 +249,43 @@ export class Endpoint<
    * @param failure The new failure schema.
    * @returns A new endpoint with updated failure schema.
    */
-  public failure<T extends z.ZodType>(failure: T): Endpoint<Q, B, P, U, S, T> {
-    return new Endpoint(this.path, this.method, this.input, { ...this.output, failure })
+  public failure<T extends z.ZodType>(failure: T): Endpoint<Q, B, P, U, S, A, T> {
+    return new Endpoint(
+      this.path,
+      this.method,
+      this.input,
+      { ...this.output, failure },
+      this.authenticated,
+      this.accessRoles,
+    )
+  }
+
+  /**
+   * Creates a new endpoint with the same path and method but different access roles.
+   * These roles are used to check if a user is authorized to the endpoint.
+   *
+   * @param roles Accepted roles for the endpoint.
+   * @returns A new endpoint with updated access roles.
+   */
+  public access(...roles: string[]): Endpoint<Q, B, P, U, S, A, F> {
+    return new Endpoint(
+      this.path,
+      this.method,
+      this.input,
+      this.output,
+      this.authenticated,
+      new Set(roles),
+    )
+  }
+
+  /**
+   * Checks if the given role is accepted by the endpoint.
+   *
+   * @param role The role to check.
+   * @returns Whether the role is accepted by the endpoint.
+   */
+  public hasAccess(role: string): boolean {
+    return this.accessRoles.has(role)
   }
 
   /**