+++ packages/api-schema/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @ivy-chess/api-schema
 
+## 0.6.2
+
+### Patch Changes
+
+- Add id to params for remove and update endpoints
+
+## 0.6.1
+
+### Patch Changes
+
+- Fix issue, where params was used instead of query
+
+## 0.6.0
+
+### Minor Changes
+
+- Added endpoints, that allow users to be removed, updated and fetched.
+
+## 0.5.0
+
+### Minor Changes
+
+- add `signOut` endpoint to auth router
+
+## 0.4.0
+
+### Minor Changes
+
+- Added support for authorization management
+
+## 0.3.0
+
+### Minor Changes
+
+- Add support for protected endpoints
+
 ## 0.2.0
 
 ### Minor Changes